+++ src/styles/global.js
@@ -1,4 +1,4 @@
-import 'typeface-source-sans-pro';
+// import 'typeface-source-sans-pro';
 import 'normalize.css';
 import 'react-colorful/dist/index.css';
 import './global-styles.less';